convert SvxTimeFormat to scoped enum
and make the numerators look more like our internal formatting codes

Reviewed-on: https://gerrit.libreoffice.org/42345
Tested-by: Jenkins <ci@libreoffice.org>

+++ sd/source/filter/ppt/pptin.cxx
@@ -1460,7 +1460,7 @@ void ImplSdPPTImport::SetHeaderFooterPageSettings( SdPage* pPage, const PptSlide
                     SvxDateFormat eDateFormat;
                     SvxTimeFormat eTimeFormat;
                     PPTFieldEntry::GetDateTime( pHFE->nAtom & 0xff, eDateFormat, eTimeFormat );
-                    rHeaderFooterSettings.meDateTimeFormat = eDateFormat | ( eTimeFormat << 4 );
+                    rHeaderFooterSettings.meDateTimeFormat = eDateFormat | ( static_cast<int>(eTimeFormat) << 4 );
                 }
                 break;
                 case 1 :
